Moved Class to Database folder
